Moved auth out of extensions
