+++ pyvo/auth/tests/data/tap/capabilities.xml
