+++ pyvo/auth/tests/setup_package.py
@@ -4,4 +4,4 @@ import os
 
 def get_package_data():
     paths = [os.path.join('data/tap', '*.xml')]
-    return {'pyvo.extensions.auth.tests': paths}
+    return {'pyvo.auth.tests': paths}
